Javadoc for team match domain objects.

+++ src/main/java/com/github/sornerol/chess/pubapi/domain/match/TeamMatchBoard.java
@@ -6,16 +6,26 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * Information about a team match board
+ */
 @Getter
 @Setter
 @ToString
 public class TeamMatchBoard {
+    /**
+     * Match scores for each player. The map key is the player username, and the value is the player's score
+     */
     @JsonProperty("board_scores")
-    Map<String, Integer> boardScores;
+    Map<String, BigDecimal> boardScores;
 
+    /**
+     * List of games for this team match board
+     */
     @JsonProperty("games")
     List<ArchiveGame> games;
 }
